% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_matomo_data.R
\name{read_matomo_data}
\alias{read_matomo_data}
\title{Read data from Matomo API}
\usage{
read_matomo_data(
  connection = NULL,
  apiModule = NULL,
  apiAction = NULL,
  idDimension = NULL,
  date = "yesterday",
  period = "day",
  format = "csv",
  processed_report = FALSE,
  idSubtable = NULL,
  pageUrl = NULL,
  pageTitle = NULL
)
}
\arguments{
\item{connection}{A connection object created by the setMatomoServer() function.}

\item{apiModule}{A character vector of an API Module from the
\href{https://developer.matomo.org/api-reference/reporting-api#api-method-list}{Reporting
API Method List}}

\item{apiAction}{A character vector of an API Action for the selected API
Module}

\item{idDimension}{A numeric vector of the Custom Dimension}

\item{date}{A date in standard ISO 8601 format YYYY-MM-DD. Magic character
vector keywords are "today" or "yesterday".}

\item{period}{Period you request the statistics for. Can be any of: day,
week, month, year or range. All reports are returned for the dates based on
the website's time zone.}

\item{format}{A character vector to define the output format as CSV, XML or
JSON}

\item{processed_report}{A logical that controls whether the method
"getProcessedReport" is used for the API Action and Module. The method will
return a human readable version of any other report, and include the
processed metrics such as conversion rate, time on site, etc. which are not
directly available in other methods.}

\item{idSubtable}{A numeric vector to identify a subtable. Only valid when
processed_report = TRUE.}

\item{pageUrl}{A character vector to identify the target url of a page. Only required for certain actions (documented in the Matomo API reference).}

\item{pageTitle}{A character vector to identify the target title of a page. Only required for certain actions (documented in the Matomo API reference).}
}
\value{
The output will be a data of the selected format
}
\description{
\code{read_matomo_data} reads data from
\href{https://developer.matomo.org/api-reference/reporting-api}{Matomo
Reporting HTTP API} for available methods.

A method is a combined API Module and API Action.
}
\examples{
conObj<-setMatomoServer()
read_matomo_data(
connection = conObj,
apiModule = "Actions", apiAction = "getPageUrls"
)

\dontrun{
read_matomo_data()
}

}
