% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMatomoServer.R
\name{setMatomoServer}
\alias{setMatomoServer}
\title{Set Connection to Matomo API}
\usage{
setMatomoServer(server = "webzh-dk", tokenString = NULL)
}
\arguments{
\item{server}{A character vector specifying which server to access. Currently it supports
the three possibilities openzh (the open data websites), webzh-dk (the data catalogue of webzh) and webzh (the websites of webzh). By default it will access webzh-dk if nothign was provided.}

\item{tokenString}{A character vector specifying which token to access. It is a voluntary argument overriding the default strings. These are token_openzh, token_webzh-dk and token_webzh.}
}
\value{
The output is a connection object, that is actually a list object containing the three needed parameters to set up the connection..
}
\description{
\code{setMatomoServer} This command is required to set the specifications to connect to the server
}
\examples{
conObj<-setMatomoServer(server="openzh")

}
